California Probate Code Section 4401 PoA
This is the statutory form specified in the CCPA regulations necessary to form a Power of Attorney.  The aim is to use this form for the authorization by the consumer to Consumer Reports to exercise their rights on their behalf under the CCPA.  The plan is to attempt two workflows, one with remote notarization   (under § 999.326(b) of the regulations) and the other without notarization (under § 999.326(a) of the regulations).  

Please note, this is just an initial draft making no changes other than transforming the content "as-is" to pure text and formating in markdown.  Iteratively, updates will be needed to wire the signatures to whatever process is selected, and to add a few words describing the more specific and limited authorization by the consumer for Consumer Reports just to exercise their rights under CCPA (probably flowing from the selection for "(F) Business operating transactions."), also whatever other adjustments may be needed for display and workflow in line with the development of the overall code and project.  I'm adding this to the commit just to keep the note and will reflect these items in project documentation, issues, or however else the project managers deem appropriate.