Improve container coordination
Previously, during initial application deployment, the "database
migration" command was repeatedly invoked until successful. Although
this deferred the execution of the application code, it also obscured
problems in the database migration. (Failures to connect to the database
were expected to resolve themselves, but failures to migrate would
require developer intervention.)

Introduce a new script designed to pause execution based solely on the
ability to connect to the database. This allows errors in the subsequent
database migration to correctly be interpreted as fatal.

Improve parity between development and production environments by
modeling the development environment's PostgreSQL connection information
according to the format provided by the Heroku service in the production
environment.